separate ttt run methods from game
